Reorganise file and categories them into different packages for better code readability
